python-interview-analyzer/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
openai==1.3.8
opencv-python==4.8.1.78
mediapipe==0.10.8
librosa==0.10.1
speechrecognition==3.10.0
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
numpy==1.25.2
scipy==1.11.4
openai-whisper==20231117

# Google Services
google-cloud-storage==2.10.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
gspread==5.12.0

# Video and Audio Processing
moviepy==1.0.3
pydub==0.25.1
ffmpeg-python==0.2.0
yt-dlp==2023.11.16

# Computer Vision and Emotion Analysis
deepface==0.0.79
fer==22.5.1
face-recognition==1.3.0
dlib==19.24.2

# Data Processing
pandas==2.1.3
pillow==10.1.0
requests==2.31.0
aiohttp==3.9.1
PyPDF2==3.0.1
python-docx==0.8.11

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.0.3
PyYAML==6.0.1

# Database and Caching
redis==5.0.1
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0

# Audio Processing
pyaudio==0.2.11
soundfile==0.12.1
webrtcvad==2.0.10

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pre-commit==3.5.0

# Monitoring and Logging
sentry-sdk[fastapi]==1.38.0
prometheus-client==0.19.0

# Deployment
gunicorn==21.2.0
docker==6.1.3
